refactor(recruitment): split notice-edit-page god-object (S1 of #141)
Pure code movement — no behavior changes.

class-ffc-recruitment-notice-edit-page.php drops from 1727 to 291 LOC.
The orchestrator class keeps register(), handle_download_csv_example(),
render(), handle_save(), handle_transition(), back_url(), and
redirect_with_notice(). The 5 section-render calls in render() and the
columns_label_map() call in handle_save() now delegate to
RecruitmentNoticeEditPageRenderer.

Two new files:

  - class-ffc-recruitment-notice-edit-page-renderer.php (1330 LOC)
    All render_*, columns_label_map, preview_status_label_map, cls_button,
    lookup_map, transitions_from, and the two inline-script emitters.
    The 6 methods called from the EditPage facade are public; the rest
    stay private (only sibling-renderer callers).

  - class-ffc-recruitment-classification-filter-manager.php (169 LOC)
    read_classification_filters() → read_filters() and
    apply_classification_filters() → apply_filters(). Bodies byte-
    identical; only the names changed per the issue spec.

PSR-4 autoloader resolves both new class names from kebab-case
filenames automatically — no loader/composer changes needed.

Verification:
  - php -l: clean on all 3 files.
  - vendor/bin/phpunit --filter Recruitment: 227/227 OK.
  - vendor/bin/phpstan analyse <3 files> --level=8: no errors.
  - Re-ran phpstan over includes/recruitment/ (34 files): no errors.
